Ajuste no envio de mensagens
Para enviar mensagem privada a um usuário, basta usar a sintaxe

usuario:mensagem desejada

O caractere separador entre USUARIO e MENSAGEM será o caractere :
