+++ README.md
@@ -4,7 +4,8 @@
 
 * [ ] Conexão com servidor, definindo host e porta (Client Side)
 * [ ] Usuário é identificado por username, então o username tem que ser único
-* [ ] Envio de mensagens públicas (Todos usuários recebem)
-* [ ] Envio de mensagens privadas (Apenas usuário específico recebe)
-* [ ] Alerta de entrada/saída de usuário no log de mensagens
+* [X] Envio de mensagens públicas (Todos usuários recebem)
+* [X] Envio de mensagens privadas (Apenas usuário específico recebe)
+* [ ] Alerta de entrada de usuário no log de mensagens
+* [X] Alerta de saída de usuário no log de mensagens
 * [ ] Lista de usuários sempre atualizada (Apenas os usuários que estão online)
